Restructure files and prepend license automatically on build
