+++ COPYRIGHT_NOTICE
@@ -16,4 +16,3 @@
 # along with pegjs-ssc-parser. If not, see <http://www.gnu.org/licenses/>.
 *
 */
-
