--- index.ts
@@ -1,3 +1,0 @@
-export { parseSSC, SyntaxError, validateSSC } from "./parser";
-export type { Parse, Level, Meta } from "./parser/types";
-export type { ValidatedSSC } from "./parser/validate";
