+++ package.json
@@ -9,9 +9,11 @@
     "url": "https://github.com/piulin/pegjs-ssc-parser"
   },
   "scripts": {
-    "build": "tsc",
-    "prepublish": "tsc",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "rimraf dist && yarn tsc && node addLicense.js",
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "format": "prettier --config .prettierrc '**/*.ts' --write",
+    "lint": "eslint src/ --ext .ts",
+    "lint:fix": "yarn lint --fix"
   },
   "keywords": [
     "SSC",
@@ -28,6 +30,13 @@
   },
   "devDependencies": {
     "@types/node": "^18.6.2",
+    "@typescript-eslint/eslint-plugin": "^5.31.0",
+    "@typescript-eslint/parser": "^5.31.0",
+    "eslint": "^8.20.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "prettier": "^2.7.1",
+    "rimraf": "^3.0.2",
     "ts-node": "^10.9.1",
     "typescript": "^4.7.4"
   }
