+++ src/index.ts
@@ -1,0 +1,3 @@
+export { parseSSC, SyntaxError, validateSSC } from './parser';
+export type { Level, Meta, Parse } from './parser/types';
+export type { ValidatedSSC } from './parser/validate';
