+++ src/parser/utils/checkSSCs.js
