+++ tsconfig.json
@@ -3,6 +3,7 @@
       "allowJs": true,
       "target": "es5",
       "module": "commonjs",
+      "rootDir": "src/",
       "lib": ["es2017", "es7", "es6", "dom"],
       "declaration": true,
       "outDir": "dist",
@@ -14,5 +15,5 @@
       "node_modules",
       "dist"
     ],
-    "include": ["**/*"]
+    "include": ["src/"]
   }
